% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{covariance_matrix_data}
\alias{covariance_matrix_data}
\title{covariance_matrix_data}
\format{
An object of class \code{data.frame} with 38436 rows and 13 columns.
}
\usage{
data("covariance_matrix_data")
}
\description{
Selected data matrix based on SNPs for calculating covariance matrix of phenotypes. 
\code{
load(file.path("data","covariance_matrix_estimated.Rdata"))
usethis::use_data(covariance_matrix_data, overwrite = TRUE)
}
}
\keyword{datasets}
